Move current player algorithm to rust library

+++ app/src/main/java/com/taavit/scorecounter/ui/GameUiState.kt
@@ -1,8 +1,8 @@
 package com.taavit.scorecounter.ui
 
 data class GameUiState(
-    val p1Score: Int = 0,
-    val p2Score: Int = 0,
+    val p1Score: Byte = 0,
+    val p2Score: Byte = 0,
     val startingPlayer: Player = Player.Player1,
     val currentServe: Player = Player.Player1,
     val isGameFinished: Boolean = false,
